requirements.txt
bleach
boto
celery
django
django-allauth
django-celery-email
django-cleanup
django-crispy-forms
django-filter
django-guardian
django-pylibmc
django-redis
django-resized
django-simple-captcha
djangorestframework
django-storages-redux
dj-database-url
markdown
mysqlclient
newrelic
psycopg2
pylibmc
pyotp
sendgrid-django
tweepy
waitress
